requirements.txt
pytorch==0.3.1
torchtext==0.2.3
argparse
tqdm